requirements.txt
flask==2.0.3
python-telegram-bot==13.7
gunicorn==20.1.0
werkzeug==2.0.3
urllib3==1.26.0